refactor: 폴더 utils => services로 변경 및 위치 이동
