+++ src/refactoring/services/cart.ts
@@ -1,4 +1,4 @@
-import { CartItem, Coupon, Product } from '../../../types';
+import { CartItem, Coupon, Product } from '../../types';
 
 export const formatPriceToLocaleString = (price: number) => {
   return price.toLocaleString();
